+++ original/list/arraylist.h
@@ -6,18 +6,18 @@
 typedef struct _tagMTL_ArrayList
 {
 	void *type;
-	int capacity;
-	int length;
+	size_t capacity;
+	size_t length;
 	void *data[];
 } *MTL_ArrayList;
 
-MTL_ArrayList new_MTL_ArrayList(int length);
+MTL_ArrayList new_MTL_ArrayList(size_t length);
 void MTL_ArrayList_release(MTL_ArrayList self);
 
-int MTL_ArrayList_length(MTL_ArrayList self);
-bool MTL_ArrayList_set(MTL_ArrayList self, int idx, void *value);
-bool MTL_ArrayList_get(MTL_ArrayList self, int idx, void **value);
-bool MTL_ArrayList_append(MTL_ArrayList self, int *idx, void *value);
+size_t MTL_ArrayList_length(MTL_ArrayList self);
+bool MTL_ArrayList_set(MTL_ArrayList self, size_t idx, void *value);
+bool MTL_ArrayList_get(MTL_ArrayList self, size_t idx, void **value);
+bool MTL_ArrayList_append(MTL_ArrayList self, size_t *idx, void *value);
 bool MTL_ArrayList_iterate(MTL_ArrayList self, bool (*func)(void *value));
 
 #endif
